Move file to app
